refactor: move the batchnorm_cudnn into TrackerExt
